split up aws and scheduler modules
