+++ src/edge/edge_server.cpp
@@ -1,4 +1,7 @@
 #include <edge/edge.h>
+
+#include <util/logging.h>
+
 #include <pistache/http.h>
 
 int main() {
