+++ src/emulator/emulator.cpp
@@ -1,6 +1,6 @@
 #include <faasm/memory.h>
 
-#include <redis/redis.h>
+#include <redis/Redis.h>
 #include <state/state.h>
 
 /**
