+++ src/redis/CMakeLists.txt
@@ -2,8 +2,8 @@ include_directories(${FAASM_INCLUDE_DIR}/redis)
 
 add_library(
         redis STATIC
-        redis.cpp
-        ${FAASM_INCLUDE_DIR}/redis/redis.h
+        Redis.cpp
+        ${FAASM_INCLUDE_DIR}/redis/Redis.h
 )
 
 target_link_libraries(
