+++ src/scheduler/CMakeLists.txt
@@ -1,9 +1,11 @@
 include_directories(${FAASM_INCLUDE_DIR}/scheduler)
 
+file(GLOB HEADERS "${FAASM_INCLUDE_DIR}/scheduler/*.h")
+
 add_library(scheduler STATIC
-        message_queue.cpp
-        scheduler.cpp
-        ${FAASM_INCLUDE_DIR}/scheduler/scheduler.h
+        MessageQueue.cpp
+        Scheduler.cpp
+        ${HEADERS}
         )
 
 target_link_libraries(
