+++ src/scheduler/MessageQueue.cpp
@@ -1,4 +1,4 @@
-#include "scheduler.h"
+#include "Scheduler.h"
 
 #include <util/logging.h>
