+++ src/upload/upload.cpp
@@ -2,7 +2,7 @@
 
 #include <util/bytes.h>
 #include <prof/prof.h>
-#include <redis/redis.h>
+#include <redis/Redis.h>
 
 
 namespace edge {
