+++ test/lib/test_sgd.cpp
@@ -6,7 +6,7 @@
 #include <faasm/matrix.h>
 
 #include <data/data.h>
-#include <redis/redis.h>
+#include <redis/Redis.h>
 #include <state/state.h>
 
 using namespace faasm;
